h264nal: make non-standard behavior explicit

+++ src/h264_common.cc
@@ -432,8 +432,10 @@ std::shared_ptr<NaluChecksum> NaluChecksum::GetNaluChecksum(
   sum = (sum >> 32) + (sum & 0xffffffff);
   // add carry
   sum += (sum >> 32);
-  // truncate to 32 bits and get one's complement
-  uint32_t answer = ~sum;
+  // truncate to 32 bits and get one's complement. sum is 64 bits wide, so
+  // dropping the top half is the truncation the comment above describes,
+  // not an accident.
+  uint32_t answer = static_cast<uint32_t>(~sum);
 
   // write sum into (generic) checksum buffer (network order)
   *(reinterpret_cast<uint32_t*>(checksum->checksum)) = htonl(answer);
